re-structure tests/common.c
