+++ tests/tomcrypt_test.h
@@ -51,6 +51,7 @@ int misc_test(void);
 int base64_test(void);
 int file_test(void);
 int multi_test(void);
+int prng_test(void);
 
 #ifdef LTC_PKCS_1
 extern const struct ltc_prng_descriptor no_prng_desc;
